sunxi: DT: update device tree files for Allwinner H3 and H5 SoCs
Update the device tree files from the Linux tree as of v4.17-rc5,
exactly Linux commit:
commit c235edcb34651023c9eb1c5c4a7cdcaf7250d02c
Author: Giulio Benetti <giulio.benetti@micronovasrl.com>
Date:   Tue Mar 13 11:16:45 2018 +0100

    ARM: dts: sun8i-h3: Add Mali node

Since the H3 and H5 are very similar (aside from the actual ARM cores),
they share most the SoC .dtsi and thus have to be updated together.
One tiny change is the removal of the "arm/" prefix from the include
path in the sun50i-h5.dtsi, which is needed because we don't share the
same sophisticated DT directory layout of Linux.
Also we need to (temporarily) fix up the BananaPi-M2 board .dts, to
maintain bisectability.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Acked-by: Maxime Ripard <maxime.ripard@bootlin.com>